Changing before/after to raster1/raster2
in SummarizeChange_functions.R
changed raster names that were 'before' and 'after'
to 'raster1' and 'raster2'

Usage:
for scenario modeling, 'raster1' is a base case and 'raster2' is one of the scenarios;
or 'raster1' is an earlier time step and 'raster2' is a later time step.
For outcome reporting, 'raster1' is before treatments, and 'raster2' is after treatments.