Writing functions to read in and process 'patch' layers (treatments/fires)
refactored clipping/cropping function into read/check crs function
and separate clipping/cropping and recalculating area functions

wrote function to read in and check crs for the patch (e.g. treatment
or fire) dataset, as well as the function to intersect it with either
a boundary or a zonal summary polygon layer, then aggregate based on
an appropriate field/attribute (e.g. region, HUC12 id, etc), and then
crop to boundary and recalculate areas (calling the other function to
do it)