Sketching last two functions in the workflow (but not written)
This commit just adds the bones of a zonal/global raster summary
function based on whichever vector layer you want
and a reminder that then I need to write a function that will call
it three times, for HUCs, HUCS/fires, and HUCS/treatments

But there's not much code in it.  Just committing so I don't leave
it in an unsaved state.