Writing functions to do zonal summaries
Both editing the old zonal calculations function with new variable names
and also writing a function that will either do a zonal or global calc
with placeholder for global still, but now calls zonal.calculations for
zonal

Also moved the zonal.calculations above the larger function or else R
can't find it

also a variety of small comments here and there.  and the script to
run/test the zonal calcs.