separate script for calculating aggregated patches
pulling out pieces from targeted effort. r and summarize change
outcome reporting.r, to just prepare the vector files for both
targeted effort and efficacy calculations.

Also edited everywhere to take spaces out of the policy target
lists (in activity list as well as the main scripts)

also corrected a bug where I was using 'end' and 'start' but those
are builtin function names, this needed edits in several scripts

also changed the name of the local folder to intermediatefiles
so that one folder within that is rasterdiffs and the other is
aggregated vectors.