Thresholded versions of efficacy calcs and data visualization
new script to do thresholded version of efficacy, and a script
to calculate the layers needed for it.

Also updates to the other scripts, including adding data visualization
to the Targeted effort script, as well as for the efficacy in the summarize
change script and the new thresholded efficacy script.